requirements.txt
google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
typer==0.12.5
pydantic==2.8.2
python-dateutil==2.9.0.post0
